requirements.txt
aioconsole==0.6.2
python-dotenv==1.0.0